Catching SecurityException and logging warning
Previously, a SecurityException during class initialization, such as when
trying to access the "user.home" system property, would cause the
ResourceBundle initialization to fail.  This failure was cached and any
subsequent attempts to access the bundle, even if it would then succeeed,
would fail.

With this change, the resource bundle will still function, but without any
direct access to source files.